Update prompt to be quicker
Change how the prompt is built so that we quickly render most of the vcs info, and asynchronously update the prompt with the git untracked/skipped status

After adding the vcs info and the custom git untracked/skipped info to the prompt I noticed that sometimes it took awhile for the prompt to be ready. To fix this I changed the vcs portion to update asynchronously. This is nice because I can quickly `cd` through dirs and get work done, but sometimes the vcs info would be stale until the async job finished and updated the prompt.

Eventually I realized 2 things:
- The long delay only happened in big git repos
- The only portion of the vcs update that was taking a long time was updating the git untracked status (ie. `git ls-files --other --exclude-standard`)

This commit changes the prompt back to adding the vcs info synchronously and only kicks off the git untracked/skipped status gathering asynchronously.

With these change:
- I can move around the shell quickly
- `git` branch, staged, unstaged and action messages (rebase, merge, etc) update quickly and are never stale
- I still get my custom git untracked/skipped status